Fixing NumberFormatException when pageNum is greater than 2,147,483,647
Numbers greater than 2,147,483,647 can also now be returned, whereas before, numbers greater than 2,147,483,647 would overflow.

+++ src/main/java/org/example/PageController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.IntStream.rangeClosed;
+import static java.util.stream.LongStream.rangeClosed;
 import static org.example.FizzBuzz.toFizzBuzzValue;
 
 /**
@@ -20,10 +20,10 @@ public class PageController {
 
     @GetMapping("/getPage")
     public Page getPage(
-            @RequestParam(value = "pageNum", defaultValue = "0") int pageNum,
+            @RequestParam(value = "pageNum", defaultValue = "0") long pageNum,
             @RequestParam(value = "pageSize", defaultValue = "100") int pageSize) {
-        int startInclusive = 1 + (pageNum * pageSize);
-        int endInclusive = (1 + pageNum) * pageSize;
+        long startInclusive = 1 + (pageNum * pageSize);
+        long endInclusive = (1 + pageNum) * pageSize;
         List<FizzBuzzValue> values
                 = rangeClosed(startInclusive, endInclusive)
                 .mapToObj(toFizzBuzzValue)
